movielens fixing

+++ cf_experiments_loop/transformations/movielens.py
@@ -4,7 +4,6 @@ import urllib
 import zipfile
 import six
 import pandas as pd
-import numpy as np
 from sklearn.model_selection import train_test_split
 import tensorflow as tf
 from signal_transformation import helpers
@@ -27,7 +26,6 @@ MOVIE_COLUMNS = [ITEM_COLUMN, TITLE_COLUMN, GENRE_COLUMN]
 
 def _transform_csv(input_path, output_path, names, skip_first, separator=","):
     """Transform csv to a regularized format.
-
     Args:
       input_path: The path of the raw csv.
       output_path: The path of the cleaned csv.
@@ -35,10 +33,8 @@ def _transform_csv(input_path, output_path, names, skip_first, separator=","):
       skip_first: Boolean of whether to skip the first line of the raw csv.
       separator: Character used to separate fields in the raw csv.
     """
-    print(six.PY2)
     if six.PY2:
         names = [six.ensure_text(n, "utf-8") for n in names]
-    print(names)
 
     with tf.io.gfile.GFile(output_path, "wb") as f_out, \
             tf.io.gfile.GFile(input_path, "rb") as f_in:
@@ -47,8 +43,6 @@ def _transform_csv(input_path, output_path, names, skip_first, separator=","):
         f_out.write(",".join(names).encode("utf-8"))
         f_out.write(b"\n")
         for i, line in enumerate(f_in):
-            print(i)
-            print(line)
             if i == 0 and skip_first:
                 continue  # ignore existing labels in the csv
 
@@ -65,40 +59,52 @@ def prepare_data(
         clear=False,
         movielens_path=None
 ):
+    raitings_file = "ratings.csv"
+    movies_file = "movies.csv"
+    working_dir = os.path.join(movielens_path, dataset_type)
+    raitings_file_path = os.path.join(movielens_path, raitings_file)
+    movies_file_path = os.path.join(movielens_path, movies_file)
+
     if clear:
-        shutil.rmtree(movielens_path, ignore_errors=True)
+        shutil.rmtree(raitings_file_path, ignore_errors=True)
+        shutil.rmtree(movies_file_path, ignore_errors=True)
+        shutil.rmtree(working_dir, ignore_errors=True)
 
     helpers.create_dir(movielens_path)
 
     data_url = "http://files.grouplens.org/datasets/movielens/"
-    raitings_file = "ratings.csv"
-    movies_file = "movies.csv"
     url = "{}{}.zip".format(data_url, dataset_type)
 
     zip_path = os.path.join(movielens_path, "{}.zip".format(dataset_type))
-    zip_path, _ = urllib.request.urlretrieve(url, zip_path)
+
+    if not os.path.exists(zip_path):
+        zip_path, _ = urllib.request.urlretrieve(url, zip_path)
 
     zipfile.ZipFile(zip_path, "r").extractall(movielens_path)
+
     os.remove(zip_path)
 
-    working_dir = os.path.join(movielens_path, dataset_type)
-
-    if dataset_type == 'ml-1m':
-
+    if dataset_type == ML_1M:
         _transform_csv(
-            input_path=os.path.join(working_dir, "ratings.csv"),
-            output_path=os.path.join(working_dir, raitings_file),
+            input_path=os.path.join(working_dir, "ratings.dat"),
+            output_path=os.path.join(movielens_path, raitings_file),
             names=RATING_COLUMNS, skip_first=False, separator="::"
         )
 
         _transform_csv(
-            input_path=os.path.join(working_dir, "movies.csv"),
-            output_path=os.path.join(working_dir, movies_file),
+            input_path=os.path.join(working_dir, "movies.dat"),
+            output_path=os.path.join(movielens_path, movies_file),
             names=MOVIE_COLUMNS, skip_first=False, separator="::"
         )
 
-        shutil.copyfile(os.path.join(working_dir, raitings_file), os.path.join(movielens_path, raitings_file))
-        shutil.copyfile(os.path.join(working_dir, movies_file), os.path.join(movielens_path, movies_file))
+        shutil.copyfile(
+            os.path.join(working_dir, raitings_file),
+            os.path.join(movielens_path, raitings_file)
+        )
+        shutil.copyfile(
+            os.path.join(working_dir, movies_file),
+            os.path.join(movielens_path, movies_file)
+        )
         tf.io.gfile.rmtree(working_dir)
 
         dataset = pd.read_csv(os.path.join(movielens_path, raitings_file))
@@ -135,18 +141,18 @@ def bpr_movielens(dataset_type=None,
     return train_data, test_data, users_number, items_number
 
 
-print(prepare_data(dataset_type='ml-1m',
+print(prepare_data(dataset_type=ML_1M,
                    clear=True,
-                   movielens_path='movielens'))
+                   movielens_path='/tmp/cf_experiments_loop/dataset/movielens'))
 
 
 # preprocessing for vaecf
 def vae_movielens(dataset_type=None,
                   clear=False,
                   movielens_path=None):
-
     train_data, test_data, users_number, items_number = prepare_data(dataset_type=dataset_type,
                                                                      clear=clear,
                                                                      movielens_path=movielens_path)
 
-    return vae_preprocess_data(train_data), vae_preprocess_data(test_data), users_number, items_number
+    return vae_preprocess_data(train_data), vae_preprocess_data(
+        test_data), users_number, items_number
